Dernier régalge des warnings

+++ Software/Source/AI/Resolution.cpp
@@ -24,7 +24,7 @@ void	Resolution::parsingMarkov(StyleSettings &style, std::vector<std::pair<char,
 	{
 	  percentage = percentage * style.getProbaFromNote(save, save) / 100;
 	  if (percentage < NOT_ENOUGH_CHANCE)
-	    (*sequence)[i].first = fixingMarkov(style, save, i + 1 == sequence->size() ? -1 : (*sequence)[i + 1].first);
+	    (*sequence)[i].first = fixingMarkov(style, save, i + static_cast<unsigned int>(1) == sequence->size() ? -1 : (*sequence)[i + 1].first);
 	}
       if (save != (*sequence)[i].first)
 	{
